Look up welcome email text by disease #2712
